refactor: 카메라설정, 액션설정컴포넌트 분할 및 UI수정
