+++ app/simulator/SimulatorPage.tsx
@@ -3,7 +3,6 @@
 import { Suspense, useEffect, useMemo, useRef, useState } from 'react';
 import { Canvas } from '@react-three/fiber';
 import { ChevronLeft, ChevronRight } from 'lucide-react';
-import { useSession } from 'next-auth/react';
 
 import { Button } from '@/components/ui/button';
 
@@ -15,6 +14,7 @@ import { useLoadingStore } from '@/stores/useLoadingStore';
 import { useLoginRedirectModalStore } from '@/stores/useLoginRedirectModalStore';
 import { useRoomSizeStore } from '@/stores/useRoomSizeStore';
 
+import CameraModeButtons from './components/buttons/CameraModeButtons';
 import CaptureCanvas from './components/capture/CaptureCanvas';
 import FurnitureController from './components/furnitures/FurnitureController';
 import FurnitureModel from './components/furnitures/FurnitureModel';
@@ -34,7 +34,6 @@ interface SimulatorPageProps {
 }
 
 export default function SimulatorPage({ mode, roomId }: SimulatorPageProps) {
-  const session = useSession();
   const { openModal } = useLoginRedirectModalStore();
   const { setLoading } = useLoadingStore();
   const [canvasCreated, setCanvasCreated] = useState(false);
@@ -251,20 +250,18 @@ export default function SimulatorPage({ mode, roomId }: SimulatorPageProps) {
             )}
           </Button>
 
-          <div className='absolute top-[20px] left-[40%] z-30'>
-            <BackgroundSelector />
+          <div className='absolute gap-2 flex top-[20px] left-[40%] z-30'>
+            <div>
+              {/* 카메라 모드 버튼*/}
+              <CameraModeButtons />
+            </div>
+            <div>
+              {/* 배경 선택 버튼들 */}
+              <BackgroundSelector />
+            </div>
           </div>
-          <div className='absolute top-[140px] right-[70px] z-30'>
-            <FurnitureController />
-          </div>
-
-          <div className='absolute top-[20px] right-[20px] z-30 flex gap-2'>
-            <Button onClick={handleSaveClick}>
-              {mode === 'edit' ? '수정하기' : '저장하기'}
-            </Button>
-            <Button variant='secondary' onClick={() => history.back()}>
-              나가기
-            </Button>
+          <div className='absolute top-[30px] right-[20px] z-30'>
+            <FurnitureController mode={mode} onSaveClick={handleSaveClick} />
           </div>
         </>
       )}
@@ -299,7 +296,6 @@ export default function SimulatorPage({ mode, roomId }: SimulatorPageProps) {
             furnitures={furnitures}
             width={roomWidth}
             height={roomHeight}
-            userId={'2'}
           />
           <div className='absolute top-0 left-0 w-full h-full opacity-0 pointer-events-none'>
             <CaptureCanvas
