+++ app/simulator/components/furnitures/FurnitureController.tsx
@@ -9,35 +9,21 @@ import type { FurnitureStoreInfo } from '@/app/types/furniture';
 
 import { useFurnitureStore } from '@/stores/useFurnitureStore';
 import { useLightingStore } from '@/stores/useLightingStore';
-import { useViewStore } from '@/stores/useViewStore';
 
+import ActionButtons from '../buttons/ActionButtons';
 import FurnitureControllerBtn from './FurnitureControllerBtn';
 import FurnitureThumbnailInfo from './FurnitureThumbnailInfo';
 import LockedScaleSlider from './LockedScaleSlider';
 import ScaleLockToggle from './ScaleLockToggle';
 import UnlockedScaleInputs from './UnlockedScaleInputs';
 
-const VALID_ANGLES = [45, 135, 225, 315];
-const TOP_VIEW_ANGLES = [0, 90, 180, 270];
-
-function convertForm3DToTop(angle3D: number): number {
-  const index3D = VALID_ANGLES.indexOf(angle3D);
-  if (index3D !== -1) {
-    return TOP_VIEW_ANGLES[index3D];
-  }
-  return 0;
-}
-
-function convertFormTopTo3D(angleTop: number): number {
-  const indexTop = TOP_VIEW_ANGLES.indexOf(angleTop);
-  if (indexTop !== -1) {
-    return VALID_ANGLES[indexTop];
-  }
-
-  return 45;
-}
-
-export default function FurnitureController() {
+export default function FurnitureController({
+  mode,
+  onSaveClick,
+}: {
+  mode: 'create' | 'edit';
+  onSaveClick: () => void;
+}) {
   const {
     furnitures,
     selectedFurnitureId,
@@ -47,11 +33,6 @@ export default function FurnitureController() {
     prevFurnitureStates,
   } = useFurnitureStore();
   const [isScaleLocked, setIsScaleLocked] = useState(false);
-
-  const angle = useViewStore((s) => s.angle);
-  const setAngle = useViewStore((s) => s.setAngle);
-  const isTopView = useViewStore((s) => s.isTopView);
-  const setIsTopView = useViewStore((s) => s.setIsTopView);
 
   const isDay = useLightingStore((state) => state.isDay);
   const setIsDay = useLightingStore((state) => state.setIsDay);
@@ -84,43 +65,10 @@ export default function FurnitureController() {
     updateSelectedFurniture,
   );
 
-  // 방 회전
-  const handleRoomRotate = () => {
-    const currentAngles = isTopView ? TOP_VIEW_ANGLES : VALID_ANGLES;
-    const currentIndex = currentAngles.indexOf(angle);
-    const nextIndex = (currentIndex + 1) % currentAngles.length;
-    const newAngle = currentAngles[nextIndex];
-    setAngle(newAngle);
-  };
-
-  // 뷰 모드 전환
-  const handleViewToggle = () => {
-    if (isTopView) {
-      const new3DAngle = convertFormTopTo3D(angle);
-      setAngle(new3DAngle);
-      setIsTopView(false);
-    } else {
-      const newTopAngle = convertForm3DToTop(angle);
-      setAngle(newTopAngle);
-      setIsTopView(true);
-    }
-  };
-
   return (
     <div className='w-[219px] px-[30px] py-[25px] rounded-[30px] bg-gradient-to-r from-white/10 to-black/20 backdrop-blur-md shadow-[0_0_15px_#00000026] flex flex-col items-center justify-center gap-4'>
-      {/* 카메라 모드 버튼 */}
-      <div className='w-full flex gap-[10px]'>
-        <FurnitureControllerBtn
-          width={14}
-          height={14}
-          icon='/assets/icons/view-mode.svg'
-          onClick={handleRoomRotate}
-        />
-        <FurnitureControllerBtn
-          text={isTopView ? '3D 뷰' : '탑 뷰'}
-          onClick={handleViewToggle}
-        />
-      </div>
+      {/* 저장, 나가기 버튼 */}
+      <ActionButtons mode={mode} onSaveClick={onSaveClick} />
 
       {/* 조명 모드 버튼 */}
       <div className='flex gap-[10px]'>
